Skip GenericForeignKey fields
The `GenericForeignKey` field does not provide a `get_internal_type`
method so when checking if it's a ForeignKey or not an AttributeError is
raised.

This adjusts the code to ignore this `AttributeError` which effectively
un-monitors the `GenericForeignKey` itself. However, it does leave the
underlying `ForeignKey` to the `ContentType` table and the primary key
storage field indexing into that table monitored. This does not enable
support for hooking on the name of the `GenericForeignKey`, but hooking
on the underlying fields that support that `GenericForeignKey` should
still be possible.

closes #42